move cmd/searcher/proto to internal/searcher/proto
