+++ internal/searcher/proto/buf.gen.yaml
