+++ internal/searcher/proto/searcher.proto
